Switch method of constructing the data proxy
Uses the new allocSharedPtr method to cache the location of the shared_ptr, then continues constructing the proxy. If a circular reference is created while in the constructor of a data proxy, then it's able to pull the shared_ptr of a proxy that's still in the process of being created out of the cache.